refactor createGuestResponse logic
